Move dmn runner
