+++ packages/online-editor/src/dmnRunner/DmnRunnerDrawerPanelContent.tsx
@@ -15,7 +15,7 @@
  */
 
 import * as React from "react";
-import { useCallback, useEffect, useMemo, useRef, useState, useLayoutEffect } from "react";
+import { useCallback, useEffect, useMemo, useRef, useState } from "react";
 import { Text, TextContent } from "@patternfly/react-core/dist/js/components/Text";
 import { Page, PageSection } from "@patternfly/react-core/dist/js/components/Page";
 import { DrawerCloseButton, DrawerPanelContent } from "@patternfly/react-core/dist/js/components/Drawer";
@@ -23,7 +23,7 @@ import { useDmnRunnerDispatch, useDmnRunnerState } from "./DmnRunnerContext";
 import { Notification } from "@kie-tools-core/notifications/dist/api";
 import { DmnRunnerMode, DmnRunnerStatus } from "./DmnRunnerStatus";
 import { TableIcon } from "@patternfly/react-icons/dist/js/icons/table-icon";
-import { useOnlineI18n } from "../../i18n";
+import { useOnlineI18n } from "../i18n";
 import {
   DecisionResult,
   DecisionResultMessage,
@@ -36,12 +36,12 @@ import {
 import { Holder } from "@kie-tools-core/react-hooks/dist/Holder";
 import { useCancelableEffect } from "@kie-tools-core/react-hooks/dist/useCancelableEffect";
 import { usePrevious } from "@kie-tools-core/react-hooks/dist/usePrevious";
-import { ErrorBoundary } from "../../reactExt/ErrorBoundary";
+import { ErrorBoundary } from "../reactExt/ErrorBoundary";
 import { EmptyState, EmptyStateBody, EmptyStateIcon } from "@patternfly/react-core/dist/js/components/EmptyState";
 import { I18nWrapped } from "@kie-tools-core/i18n/dist/react-components";
 import { ExclamationTriangleIcon } from "@patternfly/react-icons/dist/js/icons/exclamation-triangle-icon";
 import { WorkspaceFile } from "@kie-tools-core/workspaces-git-fs/dist/context/WorkspacesContext";
-import { EditorPageDockDrawerRef, PanelId } from "../EditorPageDockDrawer";
+import { EditorPageDockDrawerRef, PanelId } from "../editor/EditorPageDockDrawer";
 import { Button, ButtonVariant } from "@patternfly/react-core/dist/js/components/Button";
 import { Dropdown, DropdownItem, DropdownToggle } from "@patternfly/react-core/dist/js/components/Dropdown";
 import { Tooltip } from "@patternfly/react-core/dist/js/components/Tooltip";
@@ -50,9 +50,9 @@ import { Flex, FlexItem } from "@patternfly/react-core/dist/js/layouts/Flex";
 import { CaretDownIcon } from "@patternfly/react-icons/dist/js/icons/caret-down-icon";
 import { ToolbarItem } from "@patternfly/react-core/dist/js/components/Toolbar";
 import { DmnRunnerLoading } from "./DmnRunnerLoading";
-import { useExtendedServices } from "../../kieSandboxExtendedServices/KieSandboxExtendedServicesContext";
+import { useExtendedServices } from "../kieSandboxExtendedServices/KieSandboxExtendedServicesContext";
 import isEqual from "lodash/isEqual";
-import { deepCopyPersistenceJson, generateUuid } from "../../dmnRunnerPersistence/DmnRunnerPersistenceService";
+import { deepCopyPersistenceJson, generateUuid } from "../dmnRunnerPersistence/DmnRunnerPersistenceService";
 
 const KOGITO_JIRA_LINK = "https://issues.jboss.org/projects/KOGITO";
